Fixes, test improvements and refactorings
Fix for issue where SSWDBInstances>>unstore: with a join could delete from all tables in the join instead of just the target class's table. 
Fix for issue sending store(In:) directly to a Proxy leaving an open transaction if the store fails. 
Also test improvements and refactorings to assist with Pharo port.